Added a lot of new features
Added more API calls
	- Added `get_all_device_configurations`;
	- Added `set_high_light_mode`;
	- Added `set_noise_status`;
	- Added `set_mirror_mode`;
	- Added `set_score_board`;

Added a "reboot" method

Added a `validate_connection` method (to, for example, wait for the device to come online)

Added `find_local_device_ip`,  a way to retrieve the local device IP via the internet to (partially) automate the connection

Supported GIF sizes are: 16 x 16, 32 x 32, 64 x 64
Added `play_net_gif` to loop the GIF from a URL on the device

Added `play_local_gif` to play a GIF from the TF-card in the device (untested)

Added `play_local_gif_directory` to play the GIFs in a directory on the TF-card in the device (untested)

Added `sound_buzzer` to use the internal buzzer to play a (configurable) buzz. Defaults to what my alarm clock sounds like. Trauma? Yes.